Separate SQL description types according to purpose
